[조성진] Sprint11

+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicMessageService.java
@@ -1,11 +1,15 @@
 package com.sprint.mission.discodeit.service.basic;
 
+import com.sprint.mission.discodeit.auth.DiscodeitUserDetails;
 import com.sprint.mission.discodeit.dto.MessageDto;
 import com.sprint.mission.discodeit.dto.response.PageResponse;
 import com.sprint.mission.discodeit.entity.BinaryContent;
 import com.sprint.mission.discodeit.entity.Channel;
 import com.sprint.mission.discodeit.entity.Message;
 import com.sprint.mission.discodeit.entity.User;
+import com.sprint.mission.discodeit.event.BinaryContentCreatedEvent;
+import com.sprint.mission.discodeit.event.MessageCreatedEvent;
+import com.sprint.mission.discodeit.event.payload.MessageCreatedPayload;
 import com.sprint.mission.discodeit.exception.channel.ChannelNotFoundException;
 import com.sprint.mission.discodeit.exception.message.MessageNotFoundException;
 import com.sprint.mission.discodeit.exception.user.UserNotFoundException;
@@ -24,9 +28,11 @@ import java.util.Optional;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
-import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.authorization.AuthorizationDeniedException;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
@@ -44,6 +50,7 @@ public class BasicMessageService implements MessageService {
   private final BinaryContentStorage binaryContentStorage;
   private final PageResponseMapper pageMapper;
   private final MessageMapper messageMapper;
+  private final ApplicationEventPublisher eventPublisher;
 
   @Transactional
   @Override
@@ -60,16 +67,17 @@ public class BasicMessageService implements MessageService {
             attachment.getOriginalFilename(), attachment.getSize(), attachment.getContentType());
         binaryContentRepository.save(content);
         log.debug("[Service] 첨부파일 저장 완료: contentId={}", content.getId());
-        binaryContentStorage.put(content.getId(), attachment.getBytes());
-        log.debug("[Service] 첨부파일 물리적 저장 완료: contentId={}, contentType={}",
-            content.getId(), content.getContentType());
+        eventPublisher.publishEvent(
+            new BinaryContentCreatedEvent(content.getId(), attachment.getBytes()));
         msg.addAttachment(content);
       }
     }
 
     messageRepository.save(msg);
-    log.debug("[Service] 메세지 저장 완료: id={}", msg.getId());
-
+    eventPublisher.publishEvent(new MessageCreatedEvent(
+        new MessageCreatedPayload(channel.getId(), channel.getName(), msg.getId(), author.getId(),
+            author.getUsername(), msg.getContent())
+    ));
     log.info("[Service] 메세지 생성 성공: id={}", msg.getId());
     return toResponse(msg);
   }
@@ -94,12 +102,12 @@ public class BasicMessageService implements MessageService {
     return pageMapper.fromData(messagesDto, nextCursor, hasNext);
   }
 
-  @PreAuthorize("@messageRepository.findById(#uuid).get().author.id == authentication.principal.userDto.id")
   @Transactional
   @Override
   public MessageDto updateMessage(UUID uuid, MessageDto.MessageUpdateRequest messageReq) {
     log.debug("[Service] 메세지 수정 시작: id={}", uuid);
     Message msg = getMessageOrThrow(uuid);
+    checkMessageOwner(msg);
 
     Optional.ofNullable(messageReq.newContent()).ifPresent(msg::updateMessage);
     messageRepository.save(msg);
@@ -109,12 +117,12 @@ public class BasicMessageService implements MessageService {
     return toResponse(msg);
   }
 
-  @PreAuthorize("@messageRepository.findById(#uuid).get().author.id == authentication.principal.userDto.id")
   @Transactional
   @Override
   public void deleteMessage(UUID uuid) throws IOException {
     log.debug("[Service] 메세지 삭제 시작: id={}", uuid);
     Message msg = getMessageOrThrow(uuid);
+    checkMessageOwner(msg);
 
     for (var attachment : msg.getAttachments()) {
       binaryContentStorage.delete(attachment.getId());
@@ -123,6 +131,15 @@ public class BasicMessageService implements MessageService {
 
     messageRepository.deleteById(uuid);
     log.info("[Service] 메세지 삭제 성공: id={}", uuid);
+  }
+
+  private void checkMessageOwner(Message msg) {
+    Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+    UUID userId = ((DiscodeitUserDetails) principal).getUserDto().id();
+
+    if (msg.getAuthor().getId() != userId) {
+      throw new AuthorizationDeniedException("메세지에 권한이 없습니다");
+    }
   }
 
   private Message getMessageOrThrow(UUID messageId) {
